Support multiple activities in the info-box at the same time, make WeekMarker a type.
Now I want you to revise the rendering of the info box.
What we need to do is to make the slider be able to select not only one item at a time but several, so they all can be displayed simultaneously in the info box.
So we should have a new field this.hoveredItems being an array that contains all elements currently being hovered on by the slider.  This should be populated instead of the this.hoveredColumn, this.hoveredPoint and this.hoveredWeekMarker (delete those) with the Point, WorkoutColumn and WeekMarker (create this Interface as its not created yet).
Then in the drawInfoBox(), it should have a for loop and for each item do the current if statement it currently has inside of it, determining what type it is (for point use property type, for workoutColumn use height, for weekMarker use weekNumber) so it gets the correct render in the info-box. Because of this for loop, each item should be rendered as a separate entity in the info box.